requirements.txt
asgiref==3.4.1
cachetools==5.3.0
certifi==2021.5.30
cffi==1.15.1
charset-normalizer==2.0.3
cryptography==40.0.2
Deprecated==1.2.13
Django==3.2.5
firebase-rest-api==1.10.1
gcloud==0.17.0
google-api-core==2.11.0
google-auth==2.17.3
google-cloud-core==2.3.2
google-cloud-firestore==2.11.0
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
googlemaps==4.10.0
grpcio==1.54.0
grpcio-status==1.54.0
httplib2==0.22.0
humanfriendly==9.2
idna==3.2
jwcrypto==1.4.2
jws==0.1.3
numpy==1.24.2
oauth2client==3.0.0
pandas==1.5.3
pkce==1.0.3
proto-plus==1.22.2
protobuf==4.22.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.4.3
pyparsing==3.0.9
pyreadline==2.1
python-dateutil==2.8.2
python-jwt==4.0.0
pytz==2021.1
requests==2.28.2
requests-toolbelt==0.7.0
rsa==4.9
six==1.16.0
sqlparse==0.4.1
tzdata==2023.3
urllib3==1.26.6
whitenoise==6.4.0
wrapt==1.15.0
